Arduino Manual Control of 5DoF Robot Arm
The integration of the MPU 6050 servo Controller to the Arduino microcontroller provides an additional 16 pins for servo motors attachment hence freeing the limited PWM pins on the Arduino board for other sensors. This shows a manual control of the robot arm with the use of a potentiometer to control the various joint angles.